// src/app/models/Empleados.modelo.ts
export class ModeloEmpleado{
    
    Nombres?: string;
    Apellidos?: string;
    Telefono?: string;
    Direccion?: string;
    Email?: string;
    Edad?: string;
    fechaNacimiento?: string;
    Clave?: string;
    Sueldo?: number;
    esDirectivo?: number;
    esCliente?: number;
    empresaId?: string


}